Moved interpolation out of Ale.h
