+++ tests/ctests/IsdTests.cpp
@@ -4,7 +4,6 @@
 
 #include "gtest/gtest.h"
 
-#include "Ale.h"
 #include "Isd.h"
 #include "Util.h"
